asan: Refine code, extract asan log to error file.
